+++ regions/io/ds9/read.py
@@ -170,15 +170,16 @@ class DS9Parser:
     radius: 40.0
     """
 
-    # List of valid coordinate system
-    coordinate_systems = ['fk5', 'fk4', 'icrs', 'galactic', 'wcs', 'physical', 'image', 'ecliptic', 'J2000']
+    # List of valid coordinate system (all lowercase)
+    coordinate_systems = ['fk5', 'fk4', 'icrs', 'galactic', 'wcs',
+                          'physical', 'image', 'ecliptic', 'j2000']
     coordinate_systems += [f'wcs{letter}' for letter in string.ascii_lowercase]
 
     # Map to convert coordinate system names
     coordsys_mapping = dict(zip(coordinates.frame_transform_graph.get_names(),
                                 coordinates.frame_transform_graph.get_names()))
     coordsys_mapping['ecliptic'] = 'geocentrictrueecliptic'
-    coordsys_mapping['J2000'] = 'fk5'
+    coordsys_mapping['j2000'] = 'fk5'
 
     def __init__(self, region_string, errors='strict'):
         if errors not in ('strict', 'ignore', 'warn'):
@@ -239,8 +240,8 @@ class DS9Parser:
             return
 
         # Special case / header: parse global parameters into metadata
-        if line.lstrip()[:6] == 'global':
-            self.global_meta = self.parse_meta(line)
+        if line.lstrip()[:6].lower() == 'global':
+            self.global_meta = self.parse_meta(line.lower())
             # global_meta can specify "include=1"; never seen other options
             # used but presumably =0 means false
             self.global_meta['include'] = (False if
@@ -249,7 +250,7 @@ class DS9Parser:
             return
 
         # Try to parse the line
-        region_type_search = regex_global.search(line)
+        region_type_search = regex_global.search(line.lower())
         if region_type_search:
             include = region_type_search.groups()[0]
             region_type = region_type_search.groups()[1]
@@ -295,7 +296,8 @@ class DS9Parser:
         meta : dict
             A dictionary containing the metadata.
         """
-        keys_vals = [(x, y) for x, _, y in regex_meta.findall(meta_str.strip())]
+        # keys must be lower-casae, but data can be any case
+        keys_vals = [(x.lower(), y) for x, _, y in regex_meta.findall(meta_str.strip())]
         extra_text = regex_meta.split(meta_str.strip())[-1]
         result = {}
         for key, val in keys_vals:
@@ -453,7 +455,8 @@ class DS9RegionParser:
         # coordinate of the # symbol or end of the line (-1) if not found
         hash_or_end = self.line.find("#")
         temp = self.line[self.region_end:hash_or_end].strip(" |")
-        self.coord_str = regex_paren.sub("", temp)
+        # force all coordinate names (circle, etc) to be lower-case
+        self.coord_str = regex_paren.sub("", temp).lower()
 
         # don't want any meta_str if there is no metadata found
         if hash_or_end >= 0: